refactor(content-hub): move workspace implementations into .epicenter/workspaces
Move all workspace definitions from individual folders (journal/, posts/, etc.)
directly into .epicenter/workspaces/*.workspace.ts files. This consolidates the
workspace implementations and removes the re-export pattern.

Folders with supporting files (gmail/auth.ts, clippings/scripts/, etc.) are kept.
